% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{app_props_xml}
\alias{app_props_xml}
\title{Generate [Content_Types].xml content.}
\usage{
app_props_xml(sheet_names)
}
\arguments{
\item{sheet_files}{Character vector of worksheet filenames.}
}
\value{
Character vector of XML lines.
}
\description{
Generate [Content_Types].xml content.
}
