% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{build_sheet_xml}
\alias{build_sheet_xml}
\title{Build XML for a worksheet.}
\usage{
build_sheet_xml(df, strings_map)
}
\arguments{
\item{df}{Data frame representing the sheet (headers included).}

\item{strings_map}{Named integer vector mapping strings to indices.}
}
\value{
Character vector with XML lines.
}
\description{
Build XML for a worksheet.
}
