% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_index}
\alias{calculate_index}
\title{Calculate index}
\usage{
calculate_index(joint_ebvs, rel_weights)
}
\arguments{
\item{joint_ebvs}{df}

\item{rel_weights}{weights}
}
\value{
df
}
\description{
Calculate index
}
