% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_pedigree}
\alias{clean_pedigree}
\title{Clean pedigree}
\usage{
clean_pedigree(ped, return_stats = FALSE)
}
\arguments{
\item{ped}{raw ped}

\item{return_stats}{whether to return stats}
}
\value{
cleaned pedigree
}
\description{
Clean pedigree
}
