% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_kinship_matrix}
\alias{compute_kinship_matrix}
\title{Compute kinship matrix from a pedigree}
\usage{
compute_kinship_matrix(ped, males, females)
}
\arguments{
\item{ped}{pedigree}

\item{males}{males}

\item{females}{females}
}
\value{
list
}
\description{
Compute kinship matrix from a pedigree
}
