% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{content_types_xml}
\alias{content_types_xml}
\title{Generate [Content_Types].xml content.}
\usage{
content_types_xml(sheet_files)
}
\arguments{
\item{sheet_files}{Character vector of worksheet filenames.}
}
\value{
Character vector of XML lines.
}
\description{
Generate [Content_Types].xml content.
}
