% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{crc32_bytes}
\alias{crc32_bytes}
\title{Compute CRC32 checksum for a raw vector.}
\usage{
crc32_bytes(raw_vec)
}
\arguments{
\item{raw_vec}{Raw vector.}
}
\value{
Signed 32-bit integer representing CRC32.
}
\description{
Compute CRC32 checksum for a raw vector.
}
