% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{create_kinship_ebv_workbook}
\alias{create_kinship_ebv_workbook}
\title{Create Excel workbook for kinship and EBV results}
\usage{
create_kinship_ebv_workbook(table_results, matrix_results, full_results)
}
\arguments{
\item{table_results}{Filtered table results}

\item{matrix_results}{Matrix results with masked values}

\item{full_results}{Complete results}
}
\value{
Workbook object ready for saving
}
\description{
Create Excel workbook for kinship and EBV results
}
