% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{create_trait_inputs}
\alias{create_trait_inputs}
\title{Create dynamic trait input UI}
\usage{
create_trait_inputs(n)
}
\arguments{
\item{n}{Number of traits}
}
\value{
UI elements for trait inputs
}
\description{
Create dynamic trait input UI
}
