% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs_engine.R
\name{custom_matings}
\alias{custom_matings}
\title{Mate allocation algorithm}
\usage{
custom_matings(Candidate, Kin, max_pair_kinship = NULL, quiet = FALSE)
}
\arguments{
\item{Candidate}{Data frame with columns Indiv, Sex, n}

\item{Kin}{Kinship matrix}

\item{max_pair_kinship}{Optional numeric threshold}

\item{quiet}{Logical, suppress messages if TRUE}
}
\value{
Data frame of matings with Sire, Dam, n, Kin
}
\description{
Mate allocation algorithm
}
