% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs_engine.R
\name{custom_noffspring}
\alias{custom_noffspring}
\title{Calculate number of offspring from optimum contributions}
\usage{
custom_noffspring(Candidate, N)
}
\arguments{
\item{Candidate}{Data frame with columns Indiv, Sex, oc, optionally BV}

\item{N}{Total number of offspring}
}
\value{
Data frame with Indiv and nOff
}
\description{
Calculate number of offspring from optimum contributions
}
