% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fallback_pedigree}
\alias{fallback_pedigree}
\title{Fallback pedigree}
\usage{
fallback_pedigree(id, dadid, momid, sex, missid = "0")
}
\arguments{
\item{id}{ids}

\item{dadid}{sire ids}

\item{momid}{dam ids}

\item{sex}{sex}

\item{missid}{missing id}
}
\value{
list
}
\description{
Fallback pedigree
}
