% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{format_kinship_ebv_results}
\alias{format_kinship_ebv_results}
\title{Format kinship and EBV results for display}
\usage{
format_kinship_ebv_results(kinship_results, ebv_results, thresh)
}
\arguments{
\item{kinship_results}{Kinship analysis results}

\item{ebv_results}{EBV analysis results}

\item{thresh}{Kinship threshold}
}
\value{
Formatted results for display
}
\description{
Format kinship and EBV results for display
}
