% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{gather_sheet_strings}
\alias{gather_sheet_strings}
\title{Gather string content from a data frame (including headers) for the
shared strings table.}
\usage{
gather_sheet_strings(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Character vector of strings (with duplicates).
}
\description{
Gather string content from a data frame (including headers) for the
shared strings table.
}
