% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generate_feedback}
\alias{generate_feedback}
\title{Generate user feedback messages}
\usage{
generate_feedback(success, message, operation)
}
\arguments{
\item{success}{Whether operation was successful}

\item{message}{Success or error message}

\item{operation}{Type of operation performed}
}
\value{
Formatted feedback message
}
\description{
Generate user feedback messages
}
