% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{get_crc32_table}
\alias{get_crc32_table}
\title{Retrieve precomputed CRC32 lookup table.}
\usage{
get_crc32_table()
}
\value{
Integer vector of length 256 with CRC32 values.
}
\description{
Retrieve precomputed CRC32 lookup table.
}
