% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_add_external_resources.R
\name{golem_add_external_resources}
\alias{golem_add_external_resources}
\title{Add external resources to the Shiny application}
\usage{
golem_add_external_resources()
}
\value{
Shiny tags for inclusion in the UI
}
\description{
This function is automatically called in `app_ui()` to add external
resources such as CSS, JavaScript, favicon, and others.
}
