% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_ebvs}
\alias{process_ebvs}
\title{Process EBVs}
\usage{
process_ebvs(trait_counter, input, prefix = "")
}
\arguments{
\item{trait_counter}{n traits}

\item{input}{shiny input}

\item{prefix}{prefix}
}
\value{
list
}
\description{
Process EBVs
}
