% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_candidates}
\alias{read_candidates}
\title{Read and process candidate files}
\usage{
read_candidates(file)
}
\arguments{
\item{file}{file input}
}
\value{
list
}
\description{
Read and process candidate files
}
