% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_uploaded_table}
\alias{read_uploaded_table}
\title{Read an uploaded delimited text table}
\usage{
read_uploaded_table(file, file_type = "file")
}
\arguments{
\item{file}{Uploaded Shiny file object or file path}

\item{file_type}{Type of file for error messages}
}
\value{
data.frame
}
\description{
Read an uploaded delimited text table
}
