% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{root_relationships_xml}
\alias{root_relationships_xml}
\title{Generate [Content_Types].xml content.}
\usage{
root_relationships_xml()
}
\arguments{
\item{sheet_files}{Character vector of worksheet filenames.}
}
\value{
Character vector of XML lines.
}
\description{
Generate [Content_Types].xml content.
}
