% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optsel_fallback.R
\name{run_custom_ocs}
\alias{run_custom_ocs}
\title{Main OCS function combining all steps}
\usage{
run_custom_ocs(
  candidates_df,
  kinship_matrix,
  ebv_index,
  desired_inbreeding_rate,
  num_offspring,
  per_pair_kinship_limit = NULL
)
}
\arguments{
\item{candidates_df}{Data frame of candidate IDs and sex}

\item{kinship_matrix}{Kinship matrix}

\item{ebv_index}{Vector of estimated breeding values}

\item{desired_inbreeding_rate}{Numeric, upper limit of mean kinship}

\item{num_offspring}{Number of offspring to produce}

\item{per_pair_kinship_limit}{Optional per-pair kinship threshold}
}
\value{
List with Candidate and Mating data frames
}
\description{
Main OCS function combining all steps
}
