% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{sanitize_sheet_names}
\alias{sanitize_sheet_names}
\title{Sanitize sheet names to comply with Excel restrictions.}
\usage{
sanitize_sheet_names(names)
}
\arguments{
\item{names}{Character vector of names.}
}
\value{
Safe, unique sheet names.
}
\description{
Sanitize sheet names to comply with Excel restrictions.
}
