% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{shared_strings_xml}
\alias{shared_strings_xml}
\title{Create a simple shared strings XML document.}
\usage{
shared_strings_xml(unique_strings, total_count)
}
\arguments{
\item{unique_strings}{Character vector of unique strings.}

\item{total_count}{Total number of string occurrences in workbook.}
}
\value{
Character vector of XML lines.
}
\description{
Create a simple shared strings XML document.
}
