% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{styles_xml_content}
\alias{styles_xml_content}
\title{Minimal styles XML content.}
\usage{
styles_xml_content()
}
\value{
Character vector of XML lines.
}
\description{
Minimal styles XML content.
}
