% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{to_dos_datetime}
\alias{to_dos_datetime}
\title{Convert POSIXct timestamp to DOS date/time components.}
\usage{
to_dos_datetime(timestamp)
}
\arguments{
\item{timestamp}{POSIXct timestamp.}
}
\value{
List with `date` and `time` integer components.
}
\description{
Convert POSIXct timestamp to DOS date/time components.
}
