% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs_helpers.R
\name{validate_ocs_inputs}
\alias{validate_ocs_inputs}
\title{Validate OCS inputs before running analysis}
\usage{
validate_ocs_inputs(
  candidates_df,
  kinship_matrix,
  ebv_index,
  desired_inbreeding_rate,
  num_offspring
)
}
\arguments{
\item{candidates_df}{Candidates data frame}

\item{kinship_matrix}{Kinship matrix}

\item{ebv_index}{Breeding value indices}

\item{desired_inbreeding_rate}{Target inbreeding rate}

\item{num_offspring}{Number of offspring}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Validate OCS inputs before running analysis
}
