% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{workbook_rels_xml}
\alias{workbook_rels_xml}
\title{Minimal styles XML content.}
\usage{
workbook_rels_xml(sheet_files)
}
\value{
Character vector of XML lines.
}
\description{
Minimal styles XML content.
}
