% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{write_xlsx_pure}
\alias{write_xlsx_pure}
\title{Generate [Content_Types].xml content.}
\usage{
write_xlsx_pure(file, sheets, creator = "AlloMate")
}
\arguments{
\item{sheet_files}{Character vector of worksheet filenames.}
}
\value{
Character vector of XML lines.
}
\description{
Generate [Content_Types].xml content.
}
