% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_xlsx_writer.R
\name{xml_escape}
\alias{xml_escape}
\title{Escape XML special characters.}
\usage{
xml_escape(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Escaped character vector safe for XML content.
}
\description{
Escape XML special characters.
}
