% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{create_ocs_trait_inputs}
\alias{create_ocs_trait_inputs}
\title{Create OCS trait input UI}
\usage{
create_ocs_trait_inputs(n)
}
\arguments{
\item{n}{Number of traits}
}
\value{
UI elements for OCS trait inputs
}
\description{
Create OCS trait input UI
}
