% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optsel_fallback.R
\name{custom_candes}
\alias{custom_candes}
\title{Create candidate object similar to candes}
\usage{
custom_candes(phen, pKin, quiet = FALSE)
}
\arguments{
\item{phen}{Data frame with columns Indiv, Sex, BV, isCandidate}

\item{pKin}{Kinship matrix}

\item{quiet}{Logical, suppress messages if TRUE}
}
\value{
Object of class 'custom_candes'
}
\description{
Create candidate object similar to candes
}
