% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optsel_fallback.R
\name{custom_opticont}
\alias{custom_opticont}
\title{Custom implementation of opticont}
\usage{
custom_opticont(method, cand, con, quiet = FALSE)
}
\arguments{
\item{method}{Optimization method, e.g., "max.BV"}

\item{cand}{Object returned by custom_candes}

\item{con}{List containing constraints}

\item{quiet}{Logical, suppress messages if TRUE}
}
\value{
Object of class 'custom_opticont'
}
\description{
Custom implementation of opticont
}
