% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fallback_fixParents}
\alias{fallback_fixParents}
\title{Fallback fixParents}
\usage{
fallback_fixParents(id, sire, dam, sex, missid = "0")
}
\arguments{
\item{id}{vector of IDs}

\item{sire}{vector of sires}

\item{dam}{vector of dams}

\item{sex}{vector of sex values}

\item{missid}{missing id value}
}
\value{
data.frame
}
\description{
Fallback fixParents
}
